accumulation steps for cipher finetune / section1 plot

+++ white_box/chat_model_utils.py
@@ -37,7 +37,7 @@ MODEL_CONFIGS = {
     "llama3_8b" : {
         "model_name_or_path" : "meta-llama/Meta-Llama-3-8B-Instruct",
         "use_fast_tokenizer" : False,
-        "dtype" : "float16",
+        "dtype" : "bfloat16",
         "chat_template" : "meta-llama/Meta-Llama-3-8B-Instruct",
     },
     "llama3_8b_cais" : {
